Corrigindo componente Botao e estruturando o projeto. Fixes #23 #24
