--- ProjetoVitaHabitus/babel.config.js
@@ -1,6 +1,0 @@
-module.exports = function(api) {
-  api.cache(true);
-  return {
-    presets: ['babel-preset-expo'],
-  };
-};
