+++ assets/snack-icon.png
