+++ screens/Login.js
@@ -1,0 +1,36 @@
+import { useState } from 'react';
+import { View } from 'react-native';
+import { TextInput, Button } from 'react-native-paper';
+import Titulo from '../components/Titulo';
+import Botao from '../components/Botao';
+
+const Login = (props) => {
+  const [email, setEmail] = useState();
+  const [senha, setSenha] = useState();
+
+  const handleEntrar = () => {
+    console.log(email, senha);
+    props.onLogin();
+  };
+
+  return (
+    <View style={{ flex: 1, justifyContent: 'center', paddingHorizontal: 16 }}>
+      <Titulo>Login</Titulo>
+      <TextInput
+        label={'E-mail'}
+        keyboardType={'email-address'}
+        value={email}
+        onChangeText={(text) => setEmail(text)}
+      />
+      <TextInput
+        label={'senha'}
+        secureTextEntry={true}
+        value={senha}
+        onChangeText={(text) => setSenha(text)}
+      />
+      <Botao onPress={handleEntrar} texto={"Entrar"}/>
+    </View>
+  );
+};
+
+export default Login;
